added classes & new game logic
